Squashed commit of the following:
commit 3ffd1fb3d0960f0785ef9a2e1b81f7a52f41d54e
Author: Stuart Bowman <sbowman@mitre.org>
Date:   Wed Aug 16 13:57:02 2023 -0400

    Cleaned out non-public code/files

commit aa43d4e2dd24d2ee33f4564640910db4925cb9f3
Author: Stuart Bowman <sbowman@mitre.org>
Date:   Wed Aug 16 08:51:24 2023 -0400

    bump to 4.3.0

+++ Loader/DecodedStream.cpp
@@ -14,7 +14,7 @@
 // For further information, please contact The MITRE Corporation, Contracts Management
 // Office, 7515 Colshire Drive, McLean, VA 22102-7539, (703) 983-6000.
 //
-// 2022 The MITRE Corporation. All Rights Reserved.
+// 2023 The MITRE Corporation. All Rights Reserved.
 // ****************************************************************************
 
 #include "loader/DecodedStream.h"
@@ -315,8 +315,8 @@ bool DecodedStream::get_datum(long &s) {
 
 bool DecodedStream::get_datum(float &s) {
    string temp = get_next().get_Data();
-   static const long big = numeric_limits<float>::max();
-   static const long small = numeric_limits<float>::min();
+   static const float big = numeric_limits<float>::max();
+   static const float small = numeric_limits<float>::min();
 
    if (temp.empty() || !DecodedStream::ok())  // if temp is empty or stream not ok
    {
